src/requirements.txt
astroid==2.3.3
boto==2.49.0
boto3==1.0.0
botocore==1.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.41.0
django-bootstrap==0.2.4
django-bootstrap4==0.0.8
django-clear-cache==0.3
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-datatables-view==1.18.0
django-filter==2.2.0
django-fontawesome==1.0
django-rest-auth==0.9.5
django-ses==0.8.12
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.15.2
future==0.17.1
futures==2.2.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.1.0
pipreqs==0.4.10
psycopg2-binary==2.8.3
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
sendgrid==6.0.5
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
yarg==0.1.9
